Corrections to content folder structure
Sections are actually pages and what I thought
were pages are fieldsets.

This changes the script that builds the folder
structure so it generates YAML files for each page
with a manifest for the page names to (called
pages.yml) so their names are no baked into the
folder structure.