Add labels for use in filters and on service pages
This commit it a *proposal* of content structure. It modifies one page of the content as an example. It is backwards compatible with all current consumers of the content.

We now have the requirement to store this additional information:
- Service page: An appropriate label for the response the supplier gave to each question in the SSP
- Search results page: A label for each option in a set of filters
- Search results page: Which of the options within the filters are presented to the user

This commit adds two additional attributes for each question's `field` object:
- `filterLabel`: The label that will be shown alongside the checkbox when filtering
- `servicePageLabel`: The label that will be shown before a supplier's response on a service page

If `filterLabel` or `servicePageLabel` are omitted then the _intention_ is that the respective filter or response should not be rendered to the page. The implementation of this is left to the application consuming the content.