Change structure of question dependency
This changes the way dependencies between questions are described in the YAML,
to be more generic than just 'lot'.

This is with the aim of having a method on [content loader](https://github.com/alphagov/digitalmarketplace-utils/blob/master/dmutils/content_loader.py)
which would exclude irrelevant questions when provided with a service.