Remove null assurance type from Data Storage questions
The Data Storage questions had `assuranceApproach: null` - perhaps because for G6 we initially thought they would need assurance but it was decided that it wasn't necessary before G6 submissions went live.

Whatever the reason, there is no need for it to be there.

Current submissions tested locally without these in and all works fine.