Bump version to 0.12.3
The most recent merge did not include a version number bump, so the tagging job failed.  The latest changes could be pulled in using the commit SHA, but it would be nicer to have a version number for it.

Minor version bump as I can't see any breaking changes or new stuff.