Use option value in JSON schemas if available
When defining checkboxes or radio sets a value can be provided to each
option as well as a label. If provided this is used as the value
attribute of the option. This change updates the JSON schemas to use
that value rather than the label if the value has been provided.