Add brief criteria weighting questions
Adds question files for technical, cultural and price weightings for
three DOS lots. Questions are split between outcomes & specialists
and participants lots, since allowed value ranges are different.

"Cultural fit" is replaced with "Availability" for the participants
lot, but the field name is the same to simplify the API validation
logic.