Add descriptions to our sections
Added (markdown-flavoured) descriptions to our sections as they
currently look in the prototype.

Using the `description` key, even though `summary_page_description` key
also seems like it would be appropriate.  Since every question is on a
single page for briefs, we don't need two different description keys.

Also reworded a description that was using a dynamic variable, as it's
not something that we can do, unfortunately.