Remove `any_of` key for weighting multiquestions
Up to now, we had implemented multiquestions with the assumption that
we would have a section full of multiquestions that we needed to
submit at least one of (but not necessarily all of).

Now, we have a section with one required multiquestion (whose inner
questions are therefore all required). This means, we can simply list
a multiquestion's questions in the top-level `required` key in our
JSON schema.

By removing the `any_of` key in our multiquestions and adding an extra
check for it in the schema generator, we avoid all of the logic around
creating `anyOf`s and `dependencies` for required multiquestions.