Add `before_summary_value` question key
Required for [#118392107](https://www.pivotaltracker.com/story/show/118392107).
To add "mandatory" options to the brief view we need a way to add additional
data to question summary value. So we're adding a new key to store this.
Values from `before_summary_value` are never sent to the API, they're only
used by the content loader for display in the summary tables.