Add per-lot description support for section descriptions
We need a way to show a different description based on the brief
response lot. We've looked at possible approaches (using a different
`lot_description`, not allowing a default option and allowing to filter
on different fields apart from by lot) and decided to allow setting
description to a dictionary of `<lot_slug>: <description_for_lot>` pairs
with an optional special `default` key to avoid duplicating matching
descriptions.

We're adding a list of existing lots to the schemas to make sure any
typos are caught by YAML tests. This is a flat list of all existing
lots for now, since the schemas are shared between different frameworks.