Add ASCII slug for "G-Cloud 7 Essentials" section
This requires content loader version 1.3.0 or later to work.

URL slugs in the frontend apps are auto-generated from section names. This is generally fine, but in this case results in a non-breaking hyphen in the section slug which we really don't want, so instead we can set an ASCII slug to be used for this section.