Add tests for section and question templating
Loads all manifests with the ContentLoader and checks that all
templatable fields in questions and sections can be rendered given
a reasonable context for the manifest type.

This allows us to check that all markdown and Jinja tags in content
fields are valid and only use the expected set of context variables.