Add tests for content markdown processing
We're sticking with the one-line/multi-line rule for processing fields
with markdown: only multi-line fields are processed.

Question .question fields are used in two ways:

* As question text for questions in a section form with multiple
  questions (declaration, service submission)
* As page title for a single-question page (briefs)

We need to make sure that questions that are using markdown syntax are
always processed with markdown while questions that are used as page
titles are never interpreted as markdown (since markdown processing adds
paragraph wrapper tags and generally can't be used inside h1/title tags).

One approach to this would be to split .question field into separate
fields (using `.title` or `.name` for one-question-per-page questions).
However, in order to integrate this into existing templates we'd have
to map new field to the `.question` in the content loader to avoid
changing all question-related templates. This means that `.title` would
be essentially another syntax for "single-line question".

Instead, we add tests to check that:

* All question fields that contain markdown syntax are multi-line fields
  (tested by forcing markdown rendering on the field and comparing with
  the original value)
* All brief and brief-responses questions .question fields are
  single-line (since they are displayed as page titles)