Fix bug with jinja content being rendered with additional p tags
Previously the question_advice for the newDayRate question was
split over multiple lines, written how it naturally would be. However,
due to how jinja was rendering multiline content the final output
when rendered on the page would add additional `<p></p>` tags to
both the start and end of the question_advice. These empty `p` tags
would create a small amount of unwanted whitespace and were not
desired. The solution to fix this is to have `question_advice` all
as a single line. This is a bit nasty for reading in the editor but
hopefully not too bad a compromise as neither way is perfect.