Dynamic list schema generator to require unoptional questions
For example, if we have a dynamic list question that will produce
a list of questions with an 'evidence' field then we will require
the 'evidence' field in each item. This would mean you must
provide evidence for each item in the dynamic list.

If the 'evidence' question was optional then it would not be
required. This will be the case for questions with follow up, i.e.
we need to make sure we mark follow up questions as optional when
creating them in the future.