Add followup type null for False answer to dynamic list question
If you answer False to a dynamic list boolean question than we
should not allow an answer to the follow up question. An example of
this is that if you answer False to a `yesNo` question then it should
be invalid to provide evidence for it.

Note, we choose to not use the `additionalKeys: False` property
as this would have interfered would the `allOf` ability to have
multiple questions with follow up in a dynamic list question.