Let multiquestions return their own required fields
Getting a list of required fields for multiquestions is complicated
by the addition of followups and the fact that dynamic list / nested
multiquestions have to deal with two required fields lists: one for
the main schema and one for the nested object schema.

Allowing property schema functions to return an optional list of
required fields allows us to move that logic closer to the followup
subschema generation without having to make any changes to the
`.form_field` methods on the question classes, that could potentially
affect other parts of the application code.