Don't add lead question to the oneOf subschema required
Always listing lead-in question as required in a subschema makes the
validation fail when used with `enforce_required=False` in the API.

This wasn't a problem for brief responses, since both the lead-in and
the followup questions are nested inside one key, which is missing in
unrelated updates. But when storing them as separate top-level key we
need to be able to unset the required fields using the schema required
list.

So instead of listing the lead-in question as required in the subschemas
we add it to the main schema required list.