Splits dynamic_list from nestd_multiquestion output
Dynamic lists are serialized as an array, but other multiquestions
aren't, so it makes sense to move the array part of the schema to
the `_dynamic_list` function.

And since we can no longer dispatch to the dynamic_list property
based on the question type we need to remove it from the types
mapping.