Add schema generation for checkbox questions with followups
Since checkboxes are answered with a list of strings some of which
can have a followup we need a way to check that the question answer
list contains a given value.

There's no direct way to do this using JSON schema, but we can check
whether the list of answers contains only the values that don't have
a follow-up. So using that and the inverse of that schema we can
separate answers that only have no-followup values and answers that
contain something else, which we assume IS the follow-up value we
wanted to check for (it could be some other invalid value, in which
case the schema will still require an answer for the followup but
the parent property schema will also flag the value as invalid).