Add descriptions files for older G-Cloud frameworks
Our buyer frontend code tries to load descriptions files without a try...except, so we really want the file to be there for all frameworks, in case it needs to load it.

This hasn't been a problem up to now (and should never be a problem in production given that these old frameworks are now expired), but for certain combinations of framework statuses (when G6 or G7 is the latest 'live' G-Cloud fraemwork) in dev, preview or staging environments these missing `descriptions` files will cause an error on the homepage.