Add `locationAllLots` question to briefs_search_filters
As the different lots for DOS have different location questions, it
meant that the location filters on the list opportunities would change
depending on the selected lot.

It was decided that it would be better to have one consistent filter
that covered all cases that could be used on all pages.