Make DOS3 declaration "more G-Cloud 9"
The declaration here was originally copied from DOS2, but that was the wrong thing to do.

Improvements were made to the declaration for G-Cloud 9, including the addition of `name` keys for display in the summary table and ``assessment` keys that allow us to auto-generate the schemas used for initial assessment of declarations when the framework closes.

THE QUESTIONS HERE ARE NOT THE FINAL QUESTIONS FOR THE FRAMEWORK.
THIS IS A PLACEHOLDER FOR THE TIME BEING, TO BE REVIEWED WHEN DOS3 IS ACTUALLY ON THE CARDS.

But as a placeholder it's better to be in the most up-to-date format now, rather than risk the old format being used by mistake when the time comes.