Add statusOpenClosed field mapping for briefs search results
This adds a new mapped field, `openClosedStatus` for the award status.
It is a direct copy of what is already there as we need to use this to
continue to provide the 'open' and 'closed' filtering for search
results.

We also need to keep the original brief status in Elasticsearch so we
can extract it to provide further detail to the end user in the search
results.

The old field mapping can be removed once changes to the
`buyer-frontend` have been made to use the new `statusOpenClosed` field.