Don't copy questions that depend on changed questions
`browserAccess` needs to not be copied as it's follow up question,
`browsersSupported` is not. If the supplier had answered 'yes' to the
`browerAccess` boolean, validation expects to an answer for
`browsersSupported`. Without it, the service is in an invalid state.

Similarly, `securityTestingCCP` needs to not be copied, as it is a
follow up to `securityTestingWhat` which we no longer copy. Having and
answer for `securityTestingCCP` but not `securityTestingWhat` leaves the
service in an invalid state.

These invalid states make editing the new draft service throw 400's.
I've checked the questions that have changed and can't find any other
dependant questions that need changing.