Remove `dates.yml` files
 ## Summary
We are migrating framework lifecycle event dates/times to the API and
storing them in machine format, rather than as strings for humans. This
should decrease any risk of errors creeping in and also bring us back to
a single source of truth.

Frontends that use frameworks will need to stop loading dates from this
frameworks repo and start using dates from the API.

Bump to version 12.0.0