Add following_framework metatdata for DOS3
We're introducing a custom error page for suppliers for when
applications close and they try to follow a link that is no longer
available. Editing a service for example.

Previously they've been served a generic error page which is unhelpful.
Now they'll get an error page with more information, and some brief
details about when the next framework is expected. That's what this
content is for.