Remove changed questions from the copy-services whitelist
Questions that have changes, or changes to their followups, should
not be copied from the previous iteration. This will highlight to
the user when completing their application that a question needs
an updated answer.

If the following framework iteration does not change these questions,
they can be reinstated to the whitelist.