Revert accreditation date field to 100 char limit
The limit is 100 in digitalmarketplace-api (https://github.com/alphagov/digitalmarketplace-api/blob/f71d124/json_schemas/services-g-cloud-12-cloud-software.json#L3067).

It was originally 50, but was changed in 62d3375f1baea0bbf8b3dbb1c5b1876694868adc. This change to 50 looks like a copy-paste error to me, so reverting to 100 is the right thing to do.

I don't know, but would be surprised if increasing the limit has any negative effects.