Pull latest version of content loader and utils
Supplier frontend is already using v7.22.0, so it should be safe to update the frameworks repo as well.

Modify test to be able to strip out more complex paragraph tags. We need to do this because the previous logic was only able to strip '<p>'. However, as of https://github.com/alphagov/digitalmarketplace-content-loader/pull/77, we get more complex paragraph tags such as '<p class="govuk-body">'. The new regex should be able to detect these more complex paragraph tags and allow us to strip them.