Make it clearer that we're asking for the lab name
A 'lab' is a room in which user research is done. A 'research studio' is a building containing one or more labs. For DOS, each lab is listed as a separate service. Thus, for this question, we want to know the name of the lab. Reword to make it clearer what we want.