Add test to check whether keys have been missed from questions_to_exclude
We had an issue where we forgot to add some fields we removed from DOS5
into the `questions_to_exclude` list in `copy_services.yml`. This test
looks for names, ids, and fields that are in a framework being copied
from but not in the framework being copied to, and raises an error if
they those removed keys are not in `questions_to_exclude`.