Change brief responses availability question to date type
Although this question asks for a simple date, it is a text question. It
has been like this since it's initial commit to DOS1 [1]; I can find no
evidence explaining why this question is not a date question.

To try and make it easier for users to enter a valid date this commit
changes the question type to date for DOS5; this will not affect brief
responses from previous frameworks and since DOS5 isn't live yet it
shouldn't cause any issues with that framework. Hopefully this will not
be a problem going forward, although it will be interesting to see if
anyone has been relying on this accepting input that is not a date.

[1]: https://github.com/alphagov/digitalmarketplace-frameworks/commit/43edb2c4b64c93f33f34d6f5d33680ebb7357dc6#diff-1437d60cdf62bf2be98d263fe7b985e069f360262ee281719624776cf8618eb3