Add tests for frameworks schema scripts to pull request checks
It would be good to know if a change to framework content causes issues
with the schema generation scripts, before merging a pull request.

This commit adds GitHub actions to run the schema generation scripts
against the latest framework iterations.

Currently the latest framework iterations are hardcoded, maybe we can
change this in future to save us from forgetting to update these tests.