Fix bug in validation schema generator
Adding `{{ item }}` to dynamic list question error messages in #656
caused validation schema generation to fail because `item` was not
present in the template context.

This commit adds `item` to the context for all dynamic_list questions;
it is a bit of a hack but is the only fix I could think of.

We also add a test that tries to assert that the template fields and
`item` are not present in the schema; thanks for the suggestion @bjgill.