Add employment status question
HMRC has strongly encouraged Public Sector Bodies (PSB)  to advertise
roles as inside or outside IR35 in order to set the expectations of
suppliers from the outset. The  HMRC off-payroll rules come
[into force](https://www.gov.uk/guidance/april-2020-changes-to-off-payroll-working-for-intermediaries)
in April 2021. As such, the DOS team has determined that all published
(outcomes and specialist) opportunities will need clear indication of
the employment status of people working for the PSB.

Add a question to the DOS5 buyer journey to ask them about the
employment status of people working on their opportunity.

Add the new question to the overall summary that the buyer sees at the
end of the first question. We want to put it here because it's
important, and we want buyers to see it.

For similar reasons, put the question at the top of the display briefs
page. It's first on an outcomes brief and second on a specialists one
(after the specialists type)